ti-vision-apps: replace repo with git command
Issue: LINUXEXEC-29512

The ti-vision-apps SRC_URI is using the repo command, but it doesn't
work for WRLinux since this repo will overwrite the repo directory of
the WRLinux parent directory. Use the git command to as a replacement.
And set the SRCREV to the REL.PSDK.ANALYTICS.09.00.01.01 tag.

(LOCAL REV: NOT UPSTREAM) - Wrlinux specific

Signed-off-by: Xulin Sun <xulin.sun@windriver.com>